+++ backend/test/other_tests.py
@@ -1,0 +1,56 @@
+import unittest
+import uuid
+
+# pylint
+import app as tested_app
+
+
+class FlaskAppTests(unittest.TestCase):
+    """
+    Class for streamlining tests on the `app.py` (flask)
+    """
+
+    def setUp(self):
+        tested_app.app.config["TESTING"] = True
+        self.app = tested_app.app.test_client()
+        with tested_app.app.app_context():
+            tested_app.database.drop_all()
+            tested_app.database.create_all()
+        def test_login_new_user(self):
+        """tests basic login functionality"""
+        req = self.app.post(
+            "/signup",
+            data={
+                "email": "Aryan@gmail.com",
+                "username": "aryan",
+                "password": "fjdanksjn",
+            },
+        )
+        assert req.json
+        req = self.app.post(
+            "/login",
+            data={
+                "email": "Aryan@gmail.com",
+                "username": "aryan",
+                "password": "fjdanksjn",
+            },
+        )
+        self.assertEqual(req.json, {"Login": True})
+
+    def test_login_incorrect(self):
+        """Incorrect user login"""
+        req = self.app.post(
+            "/login",
+            data={
+                "email": "Aryan@gmail.com",
+                "username": "aryan",
+                "password": "fjdanksn",
+            },
+        )
+        self.assertEqual(
+            req.json,
+            {"Login": False, "Log": "invalid credentials"},
+        )
+
+if __name__ == "__main__":
+    unittest.main()
